+++ app-modules/community/src/Events/PostLiked.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Domains\Community\Events;
+
+use Domains\Community\Models\Like;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Queue\SerializesModels;
+
+class PostLiked
+{
+    use Dispatchable;
+    use SerializesModels;
+
+    public function __construct(
+        public Like $like,
+        public array $context = [],
+    ) {}
+}